LayoutTests/media/video-load-readyState-expected.txt

EXPECTED (video.readyState == '0') OK
EVENT(loadstart)
EXPECTED (video.readyState == '0') OK
EVENT(loadedmetadata)
EXPECTED (video.readyState == '0') OK
EVENT(loadedfirstframe)
EXPECTED (video.readyState == '1') OK
EVENT(canshowcurrentframe)
EXPECTED (video.readyState == '1') OK
EVENT(canplaythrough)
EXPECTED (video.readyState == '3') OK
END OF TEST
